% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfm_standard.R
\name{construct_fns}
\alias{construct_fns}
\title{Construct the standard list of functionals}
\usage{
construct_fns(data)
}
\arguments{
\item{data}{the data frame of observations from which to construct the functionals.}
}
\value{
a list with
\itemize{
\item listfns: the constructed functionals
\item listinputs: their inputs
\item weightsfns: the weights that they should have so that each variable
(categorical and continuous) has the influence
}
}
\description{
These functionals are:
\itemize{
\item the mean and standard deviation for each \code{numeric} or \code{integer} variable
\item Kendall's tau for each pair of \code{numeric} / \code{integer} variable;
\item the empirical proportion for each unique possibility of each variable of
type \code{character} or \code{factor}.
}
}
