% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfm.R
\name{generateSample_spatial}
\alias{generateSample_spatial}
\title{Help function to construct the matrix of the generated samples
in the case where there is a spatial constraint.}
\usage{
generateSample_spatial(
  data,
  listfns,
  listinputs,
  coordX,
  coordY,
  minDistance,
  nDraws,
  sizeSample,
  progressBar
)
}
\description{
Help function to construct the matrix of the generated samples
in the case where there is a spatial constraint.
}
\keyword{internal}
