% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfm.R
\name{generateSample_unconstrained}
\alias{generateSample_unconstrained}
\title{Help function to construct the matrix of the generated samples
in the unconstrained case.}
\usage{
generateSample_unconstrained(
  data,
  listfns,
  listinputs,
  nDraws,
  sizeSample,
  progressBar
)
}
\description{
Help function to construct the matrix of the generated samples
in the unconstrained case.
}
\keyword{internal}
