% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfm.R
\name{mfm}
\alias{mfm}
\title{Find a sample representative of a given population
by multifunctional matching}
\usage{
mfm(
  data,
  listfns,
  listinputs,
  coordX = NULL,
  coordY = NULL,
  minDistance = 0,
  nDraws,
  sizeSample,
  methodNormalization,
  weightsfns = NULL,
  progressBar = TRUE
)
}
\arguments{
\item{data}{the dataframe to use}

\item{listfns}{the list of functionals}

\item{listinputs}{the list of inputs, where
each input is a vector of columns corresponding to each functional}

\item{coordX}{the vector of X-coordinates}

\item{coordY}{the vector of Y-coordinates}

\item{minDistance}{minimum distance between two observations
of the selected sample}

\item{nDraws}{the number of random samples to generate from the population}

\item{sizeSample}{the size of the sample}

\item{methodNormalization}{the method used for normalization.
Can be \code{ecdf} for normalization by the empirical cumulative distribution
function or \code{meansd} for the normalization \eqn{(X-E[X])/sd(X)}.}

\item{weightsfns}{weights for each of the functionals. These weights can be used
to give more or less importance to some of the functionals.
It should be a vector of the same length as the \code{listfns}.}

\item{progressBar}{\code{TRUE}: display a progressbar}
}
\value{
a list with two elements
\itemize{
\item \code{sample}: the optimal set of (indices) of observations representing
best the population.
\item \code{crit}: the value of the MFM criteria.
}
}
\description{
Find a sample representative of a given population
by multifunctional matching
}
\examples{
# matching the first mean of a two-dimensional dataset
df = data.frame(rnorm(2000), rnorm(2000))
mfm(data = df, listfns = list(mean), listinputs = list(1),
    nDraws = 500, sizeSample = 10, methodNormalization = "ecdf")

}
