% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfm_sampleSize.R
\name{mfm_sampleSize}
\alias{mfm_sampleSize}
\title{Choice of the optimal sample size for the MFM algorithm}
\usage{
mfm_sampleSize(
  data,
  rangeSizeSample,
  alpha = 0.9,
  listfns = NULL,
  listinputs = NULL,
  coordX = NULL,
  coordY = NULL,
  minDistance = 0,
  nDraws = 2000,
  nReplications = 1000,
  methodNormalization = "ecdf",
  weightsfns = NULL,
  progressBar = TRUE
)
}
\arguments{
\item{data}{the dataframe to use}

\item{rangeSizeSample}{the list of the sample sizes to tested}

\item{alpha}{the percentage of reduction of the criterion}

\item{listfns}{list of functionals to be used. If \code{NULL} (the default)
then the mean, the standard deviation and Kendall's tau between each variables
are used.}

\item{listinputs}{list of inputs to the functionals.
Only used when \code{listfns} is not NULL.}

\item{coordX}{X coordinates of the points. Used for spatial data.}

\item{coordY}{Y coordinates of the points. Used for spatial data.}

\item{minDistance}{minimum distance between two observations
of the selected sample}

\item{nDraws}{number of draws to compute each MFM criterion.}

\item{nReplications}{number of replications of the MFM criterion computation.}

\item{methodNormalization}{the method used for normalization.
Can be \code{ecdf} for normalization by the empirical cumulative distribution
function or \code{meansd} for the normalization \eqn{(X-E[X])/sd(X)}.}

\item{weightsfns}{weights for each of the functionals. These weights can be used
to give more or less importance to some of the functionals.
It should be a vector of the same length as the \code{listfns}.}

\item{progressBar}{\code{TRUE}: display a progressbar}
}
\description{
Choice of the optimal sample size for the MFM algorithm
}
\examples{
variables = c("mass", "moist", "elev")
df = agridat::gartner.corn
result = mfm_sampleSize(
  data = df[,variables], rangeSizeSample = c(15,25,30),
  nDraws = 500, nReplications = 20,
  coordX = df$long, coordY = df$lat)

}
