% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfm_standard.R
\name{mfm_standard}
\alias{mfm_standard}
\title{Standard version of the MFM algorithm}
\usage{
mfm_standard(
  data,
  coordX = NULL,
  coordY = NULL,
  minDistance = 0,
  nDraws,
  sizeSample,
  methodNormalization = "ecdf"
)
}
\arguments{
\item{data}{the dataframe to use}

\item{coordX}{the vector of X-coordinates}

\item{coordY}{the vector of Y-coordinates}

\item{minDistance}{minimum distance between two observations
of the selected sample}

\item{nDraws}{the number of random samples to generate from the population}

\item{sizeSample}{the size of the sample}

\item{methodNormalization}{the method used for normalization}
}
\description{
This function uses the MFM algorithm with predefined functionals,
which are:
\itemize{
\item the mean and standard deviation for each \code{numeric} or \code{integer} variable
\item Kendall's tau for each pair of \code{numeric} / \code{integer} variable;
\item the empirical proportion for each unique possibility of each variable of
type \code{character} or \code{factor}.
}
}
\examples{
df = data.frame(X1 = rnorm(2000), X2 = rnorm(2000),
                X3 = c(rep("A",500), rep("B", 500), rep("C", 1000)))
mfm_standard(data = df, nDraws = 500, sizeSample = 10,
             coordX = rnorm(2000), coordY = rnorm(2000))

# agricultural example
df = agridat::gartner.corn
mfm_standard(data = df[,c("mass", "moist", "elev")],
             nDraws = 500, sizeSample = 10,
             coordX = df$long, coordY = df$lat)

}
